[SG-41] move: profile 패키지 생성 및 이동
