+++ feature/profile/src/main/kotlin/com/captures2024/soongan/feature/profile/route/ProfileRoute.kt
@@ -1,7 +1,7 @@
 package com.captures2024.soongan.feature.profile.route
 
 import androidx.compose.runtime.Composable
-import com.captures2024.soongan.feature.profile.ui.ProfileScreen
+import com.captures2024.soongan.feature.profile.ui.profile.ProfileScreen
 
 @Composable
 internal fun ProfileRoute() {
